todo-list-backend/requirements.txt
bigquery==0.0.35
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cors==1.0.1
cryptography==41.0.2
dacktool==0.0.7
dbstream==0.1.21
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
future==0.18.3
gevent==22.10.2
google-api-core==2.11.1
google-api-python-client==1.7.11
google-auth==2.22.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.3
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.7.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
googleauthentication==0.0.17
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
libcst==1.0.1
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.25.1
oauthlib==3.2.2
packaging==23.1
pandas==1.3.4
proto-plus==1.22.3
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tldextract==3.4.4
typing-inspect==0.9.0
typing_extensions==4.7.1
uritemplate==3.0.1
urllib3==1.26.16
Werkzeug==2.3.6
zope.event==5.0
zope.interface==6.0
